\name{ad.matrix}
\alias{ad.matrix}
\alias{ad.matrix}
\title{An adjacency matrix of a sample graph...}
\description{An adjacency matrix of a sample graph}
\details{An adjacency matrix of a random graph with some random Entre ID of 
Protein for use in example files and the vignette}
\docType{data}
\author{Yupeng Cun \email{yupeng.cun@gmail.com}}
\keyword{data}

