\name{expr}
\alias{expr}
\alias{expr}
\title{ Two expression profile matrixs and their labels}
\description{Two expression profile matrixs and thei labels}
\details{Two expression profile matrixs and thei labels of random samples.
 expr$genes is the expression profile with Entrez ID of genes; expr$y is labels of the 
expression profile. }
\docType{data}
\author{Yupeng Cun \email{yupeng.cun@gmail.com}}
\keyword{data}

