\name{Gs2}
\alias{Gs2}
\alias{Gs2}
\title{An subgraph of hub nodes}
\description{An subgraph of hub nodes, which using igraph to generate from hubs}
\details{An adjacency matrix of hubs of a random graph was used to constructed a 
sub-graph of hubs using igraph}
\docType{data}
\author{Yupeng Cun \email{yupeng.cun@gmail.com}}
\keyword{data}

