\name{netClass-package}
\alias{netClass-package}
\alias{netClass}
\docType{package}
\title{
An R package for network-Based microarray Classification

}
\description{
We implemented average gene expression of pathway (aep), pathway activitive 
classification (PAC), Hub network Classsifccation, filter via top ranked 
genes(FrSVM), smoothed t-statistic(stSVM) for two classes microarry 
classification which employed the prior information. 
}
\details{
\tabular{ll}{
Package: \tab netClass\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-27\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
%%~~ An overview of how to use the package, including the most important ~~
%%~~ functions ~~
}
\author{
Yupeng Cun

Maintainer: Yupeng Cun <yupeng.cun@gmail.com>

}
\references{
Yupeng Cun, Holger Frohlich (2013) Network and Data Integration for Biomarker Signature Disvovery via Network Smoothed T-Statistics
}
%\keyword{ biomarker discovery, microarray classification, interaction network }
%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
